debt: remove DeprecatedStyledSelect
