+++ package.json
@@ -44,7 +44,6 @@
     "currency-symbol-map": "4.0.4",
     "debug": "4.1.1",
     "dotenv": "8.2.0",
-    "downshift": "5.0.3",
     "draft-js": "0.10.5",
     "express": "4.17.1",
     "express-http-proxy": "1.6.0",
